refactor(streaming): extract spawn drain loop and session records
