+++ src/meridian/lib/streaming/__init__.py
@@ -6,7 +6,8 @@ from meridian.lib.streaming.types import ControlMessage, InjectResult
 
 if TYPE_CHECKING:
     from meridian.lib.streaming.control_socket import ControlSocketServer
-    from meridian.lib.streaming.spawn_manager import SpawnManager, SpawnSession
+    from meridian.lib.streaming.spawn_manager import SpawnManager
+    from meridian.lib.streaming.spawn_session import SpawnSession
 
 __all__ = [
     "ControlMessage",
@@ -23,7 +24,8 @@ def __getattr__(name: str) -> object:
 
         return ControlSocketServer
     if name in {"SpawnManager", "SpawnSession"}:
-        from meridian.lib.streaming.spawn_manager import SpawnManager, SpawnSession
+        from meridian.lib.streaming.spawn_manager import SpawnManager
+        from meridian.lib.streaming.spawn_session import SpawnSession
 
         return {"SpawnManager": SpawnManager, "SpawnSession": SpawnSession}[name]
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
